requirements.txt
requests
beautifulsoup4
matplotlib
pandas
python-dotenv
google-generativeai
tiktoken
kaggle
kagglehub
scikit-learn
seaborn 
streamlit